Update draft-AIwRelationalML.md
1) Is "inference" singular or plural? Does one update + analysis result in a singular inference, or multiple smaller inferences in a set?
2) are you reviewing an ML-AI pipeline, or an rML-AI pipeline?
3) Are we starting with an ML-AI pipeline, or an rML-AI pipeline? If the former, at what point (if any) do we transition from one to the other?

This is good, I think it conveys what you set out to do, but it is extremely information-dense. Your background writing paper abstracts shows, you don't waste a single character. That might make this harder to approach for a complete novice to the field. Though, given the subject matter, this is probably well past 101-level material, I'm willing to assume you're assuming some familiarity with the field.